use integer instead of float

+++ src/main/java/com/leviathanstudio/craftstudio/client/animation/ClientAnimationHandler.java
@@ -371,7 +371,7 @@ public class ClientAnimationHandler<T extends IAnimated> extends AnimationHandle
                     KeyFrame nextStretchKeyFrame = clientChannel.getNextStretchKeyFrameForBox(boxName, currentFrame);
                     int nextStretchKeyFramePosition = nextStretchKeyFrame != null ? clientChannel.getKeyFramePosition(nextStretchKeyFrame) : 0;
 
-                    float frameCount = nextStretchKeyFramePosition - prevStretchKeyFramePosition;
+                    int frameCount = nextStretchKeyFramePosition - prevStretchKeyFramePosition;
                     float strProgress = frameCount != 0 ? ((currentFrame - prevStretchKeyFramePosition) / frameCount) : 1;
                     if (strProgress > 1F || strProgress < 0F)
                         strProgress = 1F;
